VolumeManager/VolumeSettings: Fix `origin` type

+++ src/histalign/application/VolumeSettings.py
@@ -8,7 +8,7 @@ import pydantic
 
 
 class VolumeSettings(pydantic.BaseModel):
-    origin: typing.Optional[tuple[int, int, int]] = None
+    origin: typing.Optional[tuple[float, float, float]] = None
     leaning_angle: int = 0
     axes: tuple[int, int] = (0, 1)
     offset: int = 0
